backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
numpy==1.26.4
sqlalchemy
psycopg2-binary
alembic
langchain
langchain-core
langchain-community
langchain-groq
langchain-classic
groq
llama-index
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
chromadb
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.41.2
huggingface_hub
